refactor: 패키지 구성 방식 변경
- 계층형 아키텍처로 구성된 패키지를 기능 아키텍처 방식으로 패키지 구조 변경
- 모든 클래스들은 기본적으로 package-private 방식을 고수하고 타 파일에서 접근을 필요로하는 경우에만 public으로 변경
