--- src/main/java/nextstep/subway/domain/Distance.java
@@ -1,15 +1,0 @@
-package nextstep.subway.domain;
-
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class Distance {
-
-    private Long distance;
-
-    protected Distance() { }
-
-    public Distance(Long distance) {
-        this.distance = distance;
-    }
-}
