--- src/main/java/nextstep/subway/domain/Section.java
@@ -1,46 +1,0 @@
-package nextstep.subway.domain;
-
-import javax.persistence.*;
-
-@Entity
-public class Section {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "line_id")
-    private Line line;
-
-    @ManyToOne
-    @JoinColumn(name = "up_station_id")
-    private Station upStation;
-
-    @ManyToOne
-    @JoinColumn(name = "down_station_id")
-    private Station downStation;
-
-    @Embedded
-    private Distance distance;
-
-    protected Section() { }
-
-    public Section(Station upStation, Station downStation, Distance distance) {
-        this.upStation = upStation;
-        this.downStation = downStation;
-        this.distance = distance;
-    }
-
-    public void toLine(Line line) {
-        this.line = line;
-    }
-
-    public Station getUpStation() {
-        return upStation;
-    }
-
-    public Station getDownStation() {
-        return downStation;
-    }
-}
