--- src/main/java/nextstep/subway/domain/SectionRepository.java
@@ -1,6 +1,0 @@
-package nextstep.subway.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SectionRepository extends JpaRepository<Section, Long> {
-}
