--- src/main/java/nextstep/subway/dto/UpdateLineDto.java
@@ -1,19 +1,0 @@
-package nextstep.subway.dto;
-
-import nextstep.subway.domain.Line;
-
-public class UpdateLineDto {
-    private String name;
-
-    private String color;
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() { return color; }
-
-    public Line toLine() {
-        return new Line(name, color);
-    }
-}
