+++ src/main/java/nextstep/subway/section/Distance.java
@@ -1,0 +1,15 @@
+package nextstep.subway.section;
+
+import javax.persistence.Embeddable;
+
+@Embeddable
+public class Distance {
+
+    private Long distance;
+
+    protected Distance() { }
+
+    public Distance(Long distance) {
+        this.distance = distance;
+    }
+}
