+++ src/main/java/nextstep/subway/section/SectionRepository.java
@@ -1,0 +1,6 @@
+package nextstep.subway.section;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+interface SectionRepository extends JpaRepository<Section, Long> {
+}
