+++ src/main/java/nextstep/subway/station/StationService.java
@@ -1,0 +1,36 @@
+package nextstep.subway.station;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@Transactional(readOnly = true)
+class StationService {
+    private final StationRepository stationRepository;
+
+    public StationService(StationRepository stationRepository) {
+        this.stationRepository = stationRepository;
+    }
+
+    @Transactional
+    public StationResponse saveStation(StationRequest stationRequest) {
+        Station persistStation = stationRepository.save(stationRequest.toStation());
+        return StationResponse.of(persistStation);
+    }
+
+    public List<StationResponse> findAllStations() {
+        List<Station> stations = stationRepository.findAll();
+
+        return stations.stream()
+                .map(StationResponse::of)
+                .collect(Collectors.toList());
+    }
+
+    @Transactional
+    public void deleteStationById(Long id) {
+        stationRepository.deleteById(id);
+    }
+}
